Add package boilerplate
