+++ packages/phoenix-otel/IP_NOTICE
@@ -1,0 +1,5 @@
+Copyright 2024 Arize AI, Inc. All Rights Reserved.
+
+Portions of this code are patent protected by one or more of U.S. Patent Nos. 11,315,043 and 11,615,345. The patent protection includes, but is not limited to, functionality enabled by executing the portions of the code, a device or a system executing the portions of the code, a memory storing the portions of the code, etc.
+
+This software is licensed under the terms of the Elastic License 2.0 (ELv2). See LICENSE.
