+++ packages/phoenix-otel/README.md
@@ -1,0 +1,3 @@
+# arize-phoenix-otel
+
+Provides a lightweight wrapper around OpenTelemetry primitives with Phoenix-aware defaults.
