+++ packages/phoenix-otel/SECURITY
@@ -1,0 +1,27 @@
+# Security Policy
+
+## Reporting a Vulnerability
+
+Arize takes the security of our software products and services seriously, including all of the open source code repositories managed through our GitHub organizations, such as Arize-ai.
+
+Even though open source repositories are outside of the scope of our bug bounty program and therefore not eligible for bounty rewards, we will ensure that your finding gets passed along to the appropriate maintainers for remediation.
+
+If you believe you have found a security vulnerability in any Arize AI owned repository, please report it to us through coordinated disclosure.
+
+Please do not report security vulnerabilities through public GitHub issues, discussions, or pull requests.
+
+Instead, please send an email to opensource-security[@]arize.com.
+
+Please include as much of the information listed below as you can to help us better understand and resolve the issue:
+
+The type of issue (e.g., buffer overflow, SQL injection, or cross-site scripting)
+Full paths of source file(s) related to the manifestation of the issue
+The location of the affected source code (tag/branch/commit or direct URL)
+Any special configuration required to reproduce the issue
+Step-by-step instructions to reproduce the issue
+Proof-of-concept or exploit code (if possible)
+Impact of the issue, including how an attacker might exploit the issue
+This information will help us triage your report more quickly.
+
+Policy
+See Arize AI's [trust center](https://arize.com/trust-center/)
